+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## 1.0.0
 * Add delivery details to delivery record
+* Add Delivery Order Restrictions
 * Add document links to delivery record
 * Add edd_email_template_tags
 * Add file edit links
@@ -20,8 +21,8 @@
 * Initial code release 
 * PDF counter for small ones
 * Relates docs to delivery record
-* Remove Process plugin action link
 * Remove admin only restriction
 * Remove mass processor
+* Remove Process plugin action link
 * Set plugin URL to aihr.us
 * Verbiage updates